requirements.txt


torch>=2.1.0
torchvision>=0.16.0

diffusers>=0.30.0
transformers>=4.37.0
accelerate>=0.25.0
ftfy>=6.1.0  

peft>=0.8.2
bitsandbytes>=0.41.0  
wandb>=0.17.0

opencv-python>=4.8.1.78
Pillow>=10.0.0

numpy>=1.24.4
tqdm>=4.66.1

safetensors>=0.4.2
huggingface-hub>=0.20.0

einops>=0.7.0

dotenv

scipy>=1.7.0          
scikit-image>=0.19.0   
lpips